[FLINK-19467][runtime / state backends] Add HashMapStateBackend and EmbeddedRocksDBStateBackend
