+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/hashmap/HashMapStateBackendFactory.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.runtime.state.hashmap;
+
+import org.apache.flink.configuration.IllegalConfigurationException;
+import org.apache.flink.configuration.ReadableConfig;
+import org.apache.flink.runtime.state.StateBackendFactory;
+
+/**
+ * A factory that creates an {@link HashMapStateBackend} from a configuration.
+ */
+public class HashMapStateBackendFactory implements StateBackendFactory<HashMapStateBackend> {
+
+	@Override
+	public HashMapStateBackend createFromConfig(ReadableConfig config, ClassLoader classLoader) throws IllegalConfigurationException {
+		return new HashMapStateBackend().configure(config, classLoader);
+	}
+}
