+++ flink-state-backends/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/RocksDBStateBackend.java
@@ -18,17 +18,13 @@
 package org.apache.flink.contrib.streaming.state;
 
 import org.apache.flink.annotation.VisibleForTesting;
-import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.configuration.CheckpointingOptions;
-import org.apache.flink.configuration.IllegalConfigurationException;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.core.fs.CloseableRegistry;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.runtime.execution.Environment;
-import org.apache.flink.runtime.memory.OpaqueMemoryResource;
 import org.apache.flink.runtime.query.TaskKvStateRegistry;
 import org.apache.flink.runtime.state.AbstractKeyedStateBackend;
 import org.apache.flink.runtime.state.AbstractManagedMemoryStateBackend;
@@ -37,50 +33,45 @@ import org.apache.flink.runtime.state.CheckpointStorage;
 import org.apache.flink.runtime.state.CheckpointStorageAccess;
 import org.apache.flink.runtime.state.CompletedCheckpointStorageLocation;
 import org.apache.flink.runtime.state.ConfigurableStateBackend;
-import org.apache.flink.runtime.state.DefaultOperatorStateBackendBuilder;
 import org.apache.flink.runtime.state.KeyGroupRange;
 import org.apache.flink.runtime.state.KeyedStateHandle;
-import org.apache.flink.runtime.state.LocalRecoveryConfig;
 import org.apache.flink.runtime.state.OperatorStateBackend;
 import org.apache.flink.runtime.state.OperatorStateHandle;
 import org.apache.flink.runtime.state.StateBackend;
-import org.apache.flink.runtime.state.StreamCompressionDecorator;
 import org.apache.flink.runtime.state.filesystem.FsStateBackend;
 import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
-import org.apache.flink.util.AbstractID;
-import org.apache.flink.util.DynamicCodeLoadingException;
-import org.apache.flink.util.FileUtils;
-import org.apache.flink.util.FlinkRuntimeException;
-import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.TernaryBoolean;
 
-import org.rocksdb.NativeLibraryLoader;
-import org.rocksdb.RocksDB;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-import java.io.File;
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
-import java.util.Random;
-import java.util.UUID;
 
-import static org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions.WRITE_BATCH_SIZE;
-import static org.apache.flink.contrib.streaming.state.RocksDBOptions.CHECKPOINT_TRANSFER_THREAD_NUM;
-import static org.apache.flink.contrib.streaming.state.RocksDBOptions.TIMER_SERVICE_FACTORY;
-import static org.apache.flink.util.Preconditions.checkArgument;
 import static org.apache.flink.util.Preconditions.checkNotNull;
 
 /**
- * A State Backend that stores its state in {@code RocksDB}. This state backend can
+ *<b>IMPORTANT</b> {@link RocksDBStateBackend} is deprecated in favor of {@link EmbeddedRocksDBStateBackend}.
+ * and {@link org.apache.flink.runtime.state.storage.FileSystemCheckpointStorage}. This change does
+ * not affect the runtime characteristics of your Jobs and is simply an API change to help better
+ * communicate the ways Flink separates local state storage from fault tolerance. Jobs can be upgraded
+ * without loss of state. If configuring your state backend via the {@code StreamExecutionEnvironment}
+ * please make the following changes.
+ *
+ * <pre>{@code
+ * 		StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
+ * 		env.setStateBackend(new EmbeddedRocksDBStateBackend());
+ * 		env.getCheckpointConfig().setCheckpointStorage("hdfs://checkpoints");
+ * }</pre>
+ *
+ *<p>If you are configuring your state backend via the {@code flink-conf.yaml} no changes are required.
+ *
+ *<p>A State Backend that stores its state in {@code RocksDB}. This state backend can
  * store very large state that exceeds memory and spills to disk.
  *
  * <p>All key/value state (including windows) is stored in the key/value index of RocksDB.
@@ -92,6 +83,7 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
  * using the methods {@link #setPredefinedOptions(PredefinedOptions)} and
  * {@link #setRocksDBOptions(RocksDBOptionsFactory)}.
  */
+@Deprecated
 public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 		implements CheckpointStorage, ConfigurableStateBackend {
 
@@ -107,70 +99,14 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 
 	private static final Logger LOG = LoggerFactory.getLogger(RocksDBStateBackend.class);
 
-	/** The number of (re)tries for loading the RocksDB JNI library. */
-	private static final int ROCKSDB_LIB_LOADING_ATTEMPTS = 3;
-
-	/** Flag whether the native library has been loaded. */
-	private static boolean rocksDbInitialized = false;
-
-	private static final int UNDEFINED_NUMBER_OF_TRANSFER_THREADS = -1;
-	private static final long UNDEFINED_WRITE_BATCH_SIZE = -1;
-
 	// ------------------------------------------------------------------------
 
 	// -- configuration values, set in the application / configuration
 
+	private final EmbeddedRocksDBStateBackend rocksDBStateBackend;
+
 	/** The checkpoint storage that we use for creating checkpoint streams. */
 	private final StateBackend checkpointStreamBackend;
-
-	/** Base paths for RocksDB directory, as configured.
-	 * Null if not yet set, in which case the configuration values will be used.
-	 * The configuration defaults to the TaskManager's temp directories. */
-	@Nullable
-	private File[] localRocksDbDirectories;
-
-	/** The pre-configured option settings. */
-	@Nullable
-	private PredefinedOptions predefinedOptions;
-
-	/** The options factory to create the RocksDB options in the cluster. */
-	@Nullable
-	private RocksDBOptionsFactory rocksDbOptionsFactory;
-
-	/** This determines if incremental checkpointing is enabled. */
-	private final TernaryBoolean enableIncrementalCheckpointing;
-
-	/** Thread number used to transfer (download and upload) state, default value: 1. */
-	private int numberOfTransferThreads;
-
-	/** The configuration for memory settings (pool sizes, etc.). */
-	private final RocksDBMemoryConfiguration memoryConfiguration;
-
-	/** This determines the type of priority queue state. */
-	@Nullable
-	private PriorityQueueStateType priorityQueueStateType;
-
-	/** The default rocksdb metrics options. */
-	private final RocksDBNativeMetricOptions defaultMetricOptions;
-
-	// -- runtime values, set on TaskManager when initializing / using the backend
-
-	/** Base paths for RocksDB directory, as initialized. */
-	private transient File[] initializedDbBasePaths;
-
-	/** JobID for uniquifying backup paths. */
-	private transient JobID jobId;
-
-	/** The index of the next directory to be used from {@link #initializedDbBasePaths}.*/
-	private transient int nextDirectory;
-
-	/** Whether we already lazily initialized our local storage directories. */
-	private transient boolean isInitialized;
-
-	/**
-	 * Max consumed memory size for one batch in {@link RocksDBWriteBatchWrapper}, default value 2mb.
-	 */
-	private long writeBatchSize;
 
 	// ------------------------------------------------------------------------
 
@@ -246,7 +182,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * here where the snapshots from RocksDB would be stored.
 	 *
 	 * <p>The snapshots of the RocksDB state will be stored using the given backend's
-	 * {@link StateBackend#createCheckpointStorage(JobID)}.
+	 * {@link CheckpointStorage#createCheckpointStorage(JobID)}.
 	 *
 	 * @param checkpointStreamBackend The backend write the checkpoint streams to.
 	 */
@@ -270,13 +206,9 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 			throw new IllegalStateException("RocksDBStateBackend can only checkpoint"
 				+ "to state backends that also implement CheckpointStorage.");
 		}
-
 		this.checkpointStreamBackend = checkNotNull(checkpointStreamBackend);
-		this.enableIncrementalCheckpointing = enableIncrementalCheckpointing;
-		this.numberOfTransferThreads = UNDEFINED_NUMBER_OF_TRANSFER_THREADS;
-		this.defaultMetricOptions = new RocksDBNativeMetricOptions();
-		this.memoryConfiguration = new RocksDBMemoryConfiguration();
-		this.writeBatchSize = UNDEFINED_WRITE_BATCH_SIZE;
+		this.rocksDBStateBackend = new EmbeddedRocksDBStateBackend(enableIncrementalCheckpointing);
+		this.rocksDBStateBackend.setLogger(LOG);
 	}
 
 	/**
@@ -308,69 +240,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 		this.checkpointStreamBackend = originalStreamBackend instanceof ConfigurableStateBackend ?
 				((ConfigurableStateBackend) originalStreamBackend).configure(config, classLoader) :
 				originalStreamBackend;
-
-		// configure incremental checkpoints
-		this.enableIncrementalCheckpointing = original.enableIncrementalCheckpointing.resolveUndefined(
-			config.get(CheckpointingOptions.INCREMENTAL_CHECKPOINTS));
-
-		if (original.numberOfTransferThreads == UNDEFINED_NUMBER_OF_TRANSFER_THREADS) {
-			this.numberOfTransferThreads = config.get(CHECKPOINT_TRANSFER_THREAD_NUM);
-		} else {
-			this.numberOfTransferThreads = original.numberOfTransferThreads;
-		}
-
-		if (original.writeBatchSize == UNDEFINED_WRITE_BATCH_SIZE) {
-			this.writeBatchSize = config.get(WRITE_BATCH_SIZE).getBytes();
-		} else {
-			this.writeBatchSize = original.writeBatchSize;
-		}
-
-		this.memoryConfiguration = RocksDBMemoryConfiguration.fromOtherAndConfiguration(original.memoryConfiguration, config);
-		this.memoryConfiguration.validate();
-
-		if (null == original.priorityQueueStateType) {
-			this.priorityQueueStateType = config.get(TIMER_SERVICE_FACTORY);
-		} else {
-			this.priorityQueueStateType = original.priorityQueueStateType;
-		}
-
-		// configure local directories
-		if (original.localRocksDbDirectories != null) {
-			this.localRocksDbDirectories = original.localRocksDbDirectories;
-		}
-		else {
-			final String rocksdbLocalPaths = config.get(RocksDBOptions.LOCAL_DIRECTORIES);
-			if (rocksdbLocalPaths != null) {
-				String[] directories = rocksdbLocalPaths.split(",|" + File.pathSeparator);
-
-				try {
-					setDbStoragePaths(directories);
-				}
-				catch (IllegalArgumentException e) {
-					throw new IllegalConfigurationException("Invalid configuration for RocksDB state " +
-							"backend's local storage directories: " + e.getMessage(), e);
-				}
-			}
-		}
-
-		// configure metric options
-		this.defaultMetricOptions = RocksDBNativeMetricOptions.fromConfig(config);
-
-		// configure RocksDB predefined options
-		this.predefinedOptions = original.predefinedOptions == null ?
-			PredefinedOptions.valueOf(config.get(RocksDBOptions.PREDEFINED_OPTIONS)) : original.predefinedOptions;
-		LOG.info("Using predefined options: {}.", predefinedOptions.name());
-
-		// configure RocksDB options factory
-		try {
-			rocksDbOptionsFactory = configureOptionsFactory(
-				original.rocksDbOptionsFactory,
-				config.get(RocksDBOptions.OPTIONS_FACTORY),
-				config,
-				classLoader);
-		} catch (DynamicCodeLoadingException e) {
-			throw new FlinkRuntimeException(e);
-		}
+		this.rocksDBStateBackend = original.rocksDBStateBackend.configure(config, classLoader);
 	}
 
 	// ------------------------------------------------------------------------
@@ -406,59 +276,6 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 		return checkpointStreamBackend;
 	}
 
-	private void lazyInitializeForJob(
-			Environment env,
-			@SuppressWarnings("unused") String operatorIdentifier) throws IOException {
-
-		if (isInitialized) {
-			return;
-		}
-
-		this.jobId = env.getJobID();
-
-		// initialize the paths where the local RocksDB files should be stored
-		if (localRocksDbDirectories == null) {
-			// initialize from the temp directories
-			initializedDbBasePaths = env.getIOManager().getSpillingDirectories();
-		}
-		else {
-			List<File> dirs = new ArrayList<>(localRocksDbDirectories.length);
-			StringBuilder errorMessage = new StringBuilder();
-
-			for (File f : localRocksDbDirectories) {
-				File testDir = new File(f, UUID.randomUUID().toString());
-				if (!testDir.mkdirs()) {
-					String msg = "Local DB files directory '" + f
-							+ "' does not exist and cannot be created. ";
-					LOG.error(msg);
-					errorMessage.append(msg);
-				} else {
-					dirs.add(f);
-				}
-				//noinspection ResultOfMethodCallIgnored
-				testDir.delete();
-			}
-
-			if (dirs.isEmpty()) {
-				throw new IOException("No local storage directories available. " + errorMessage);
-			} else {
-				initializedDbBasePaths = dirs.toArray(new File[dirs.size()]);
-			}
-		}
-
-		nextDirectory = new Random().nextInt(initializedDbBasePaths.length);
-
-		isInitialized = true;
-	}
-
-	private File getNextStoragePath() {
-		int ni = nextDirectory + 1;
-		ni = ni >= initializedDbBasePaths.length ? 0 : ni;
-		nextDirectory = ni;
-
-		return initializedDbBasePaths[ni];
-	}
-
 	// ------------------------------------------------------------------------
 	//  Checkpoint initialization and persistent storage
 	// ------------------------------------------------------------------------
@@ -490,20 +307,17 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 			MetricGroup metricGroup,
 			@Nonnull Collection<KeyedStateHandle> stateHandles,
 			CloseableRegistry cancelStreamRegistry) throws IOException {
-		return createKeyedStateBackend(
-			env,
-			jobID,
-			operatorIdentifier,
-			keySerializer,
-			numberOfKeyGroups,
-			keyGroupRange,
-			kvStateRegistry,
-			ttlTimeProvider,
-			metricGroup,
-			stateHandles,
-			cancelStreamRegistry,
-			1.0
-		);
+		return rocksDBStateBackend.createKeyedStateBackend(env,
+				jobID,
+				operatorIdentifier,
+				keySerializer,
+				numberOfKeyGroups,
+				keyGroupRange,
+				kvStateRegistry,
+				ttlTimeProvider,
+				metricGroup,
+				stateHandles,
+				cancelStreamRegistry);
 	}
 
 	@Override
@@ -520,57 +334,19 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 		@Nonnull Collection<KeyedStateHandle> stateHandles,
 		CloseableRegistry cancelStreamRegistry,
 		double managedMemoryFraction) throws IOException {
-
-		// first, make sure that the RocksDB JNI library is loaded
-		// we do this explicitly here to have better error handling
-		String tempDir = env.getTaskManagerInfo().getTmpDirectories()[0];
-		ensureRocksDBIsLoaded(tempDir);
-
-		// replace all characters that are not legal for filenames with underscore
-		String fileCompatibleIdentifier = operatorIdentifier.replaceAll("[^a-zA-Z0-9\\-]", "_");
-
-		lazyInitializeForJob(env, fileCompatibleIdentifier);
-
-		File instanceBasePath = new File(
-			getNextStoragePath(),
-			"job_" + jobId + "_op_" + fileCompatibleIdentifier + "_uuid_" + UUID.randomUUID());
-
-		LocalRecoveryConfig localRecoveryConfig =
-			env.getTaskStateManager().createLocalRecoveryConfig();
-
-		final OpaqueMemoryResource<RocksDBSharedResources> sharedResources = RocksDBOperationUtils
-				.allocateSharedCachesIfConfigured(memoryConfiguration, env.getMemoryManager(), managedMemoryFraction, LOG);
-		if (sharedResources != null) {
-			LOG.info("Obtained shared RocksDB cache of size {} bytes", sharedResources.getSize());
-		}
-		final RocksDBResourceContainer resourceContainer = createOptionsAndResourceContainer(sharedResources);
-
-		ExecutionConfig executionConfig = env.getExecutionConfig();
-		StreamCompressionDecorator keyGroupCompressionDecorator = getCompressionDecorator(executionConfig);
-		RocksDBKeyedStateBackendBuilder<K> builder = new RocksDBKeyedStateBackendBuilder<>(
-			operatorIdentifier,
-			env.getUserCodeClassLoader().asClassLoader(),
-			instanceBasePath,
-			resourceContainer,
-			stateName -> resourceContainer.getColumnOptions(),
-			kvStateRegistry,
-			keySerializer,
-			numberOfKeyGroups,
-			keyGroupRange,
-			executionConfig,
-			localRecoveryConfig,
-			getPriorityQueueStateType(),
-			ttlTimeProvider,
-			metricGroup,
-			stateHandles,
-			keyGroupCompressionDecorator,
-			cancelStreamRegistry
-		)
-			.setEnableIncrementalCheckpointing(isIncrementalCheckpointsEnabled())
-			.setNumberOfTransferingThreads(getNumberOfTransferThreads())
-			.setNativeMetricOptions(resourceContainer.getMemoryWatcherOptions(defaultMetricOptions))
-			.setWriteBatchSize(getWriteBatchSize());
-		return builder.build();
+		return rocksDBStateBackend.createKeyedStateBackend(
+				env,
+				jobID,
+				operatorIdentifier,
+				keySerializer,
+				numberOfKeyGroups,
+				keyGroupRange,
+				kvStateRegistry,
+				ttlTimeProvider,
+				metricGroup,
+				stateHandles,
+				cancelStreamRegistry,
+				managedMemoryFraction);
 	}
 
 	@Override
@@ -579,74 +355,16 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 		String operatorIdentifier,
 		@Nonnull Collection<OperatorStateHandle> stateHandles,
 		CloseableRegistry cancelStreamRegistry) throws Exception {
-
-		//the default for RocksDB; eventually there can be a operator state backend based on RocksDB, too.
-		final boolean asyncSnapshots = true;
-		return new DefaultOperatorStateBackendBuilder(
-			env.getUserCodeClassLoader().asClassLoader(),
-			env.getExecutionConfig(),
-			asyncSnapshots,
-			stateHandles,
-			cancelStreamRegistry).build();
-	}
-
-	private RocksDBOptionsFactory configureOptionsFactory(
-			@Nullable RocksDBOptionsFactory originalOptionsFactory,
-			String factoryClassName,
-			ReadableConfig config,
-			ClassLoader classLoader) throws DynamicCodeLoadingException {
-
-		if (originalOptionsFactory != null) {
-			if (originalOptionsFactory instanceof ConfigurableRocksDBOptionsFactory) {
-				originalOptionsFactory = ((ConfigurableRocksDBOptionsFactory) originalOptionsFactory).configure(config);
-			}
-			LOG.info("Using application-defined options factory: {}.", originalOptionsFactory);
-
-			return originalOptionsFactory;
-		}
-
-		// if using DefaultConfigurableOptionsFactory by default, we could avoid reflection to speed up.
-		if (factoryClassName.equalsIgnoreCase(DefaultConfigurableOptionsFactory.class.getName())) {
-			DefaultConfigurableOptionsFactory optionsFactory = new DefaultConfigurableOptionsFactory();
-			optionsFactory.configure(config);
-			LOG.info("Using default options factory: {}.", optionsFactory);
-
-			return optionsFactory;
-		} else {
-			try {
-				@SuppressWarnings("rawtypes")
-				Class<? extends RocksDBOptionsFactory> clazz =
-					Class.forName(factoryClassName, false, classLoader)
-						.asSubclass(RocksDBOptionsFactory.class);
-
-				RocksDBOptionsFactory optionsFactory = clazz.newInstance();
-				if (optionsFactory instanceof ConfigurableRocksDBOptionsFactory) {
-					optionsFactory = ((ConfigurableRocksDBOptionsFactory) optionsFactory).configure(config);
-				}
-				LOG.info("Using configured options factory: {}.", optionsFactory);
-
-				return optionsFactory;
-			} catch (ClassNotFoundException e) {
-				throw new DynamicCodeLoadingException(
-					"Cannot find configured options factory class: " + factoryClassName, e);
-			} catch (ClassCastException | InstantiationException | IllegalAccessException e) {
-				throw new DynamicCodeLoadingException("The class configured under '" +
-					RocksDBOptions.OPTIONS_FACTORY.key() + "' is not a valid options factory (" +
-					factoryClassName + ')', e);
-			}
-		}
+		return rocksDBStateBackend.createOperatorStateBackend(
+				env,
+				operatorIdentifier,
+				stateHandles,
+				cancelStreamRegistry);
 	}
 
 	// ------------------------------------------------------------------------
 	//  Parameters
 	// ------------------------------------------------------------------------
-
-	/**
-	 * Gets the memory configuration object, which offers settings to control RocksDB's memory usage.
-	 */
-	public RocksDBMemoryConfiguration getMemoryConfiguration() {
-		return memoryConfiguration;
-	}
 
 	/**
 	 * Sets the path where the RocksDB local database files should be stored on the local
@@ -680,53 +398,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @param paths The paths across which the local RocksDB database files will be spread.
 	 */
 	public void setDbStoragePaths(String... paths) {
-		if (paths == null) {
-			localRocksDbDirectories = null;
-		}
-		else if (paths.length == 0) {
-			throw new IllegalArgumentException("empty paths");
-		}
-		else {
-			File[] pp = new File[paths.length];
-
-			for (int i = 0; i < paths.length; i++) {
-				final String rawPath = paths[i];
-				final String path;
-
-				if (rawPath == null) {
-					throw new IllegalArgumentException("null path");
-				}
-				else {
-					// we need this for backwards compatibility, to allow URIs like 'file:///'...
-					URI uri = null;
-					try {
-						uri = new Path(rawPath).toUri();
-					}
-					catch (Exception e) {
-						// cannot parse as a path
-					}
-
-					if (uri != null && uri.getScheme() != null) {
-						if ("file".equalsIgnoreCase(uri.getScheme())) {
-							path = uri.getPath();
-						}
-						else {
-							throw new IllegalArgumentException("Path " + rawPath + " has a non-local scheme");
-						}
-					}
-					else {
-						path = rawPath;
-					}
-				}
-
-				pp[i] = new File(path);
-				if (!pp[i].isAbsolute()) {
-					throw new IllegalArgumentException("Relative paths are not supported");
-				}
-			}
-
-			localRocksDbDirectories = pp;
-		}
+		rocksDBStateBackend.setDbStoragePaths(paths);
 	}
 
 	/**
@@ -741,22 +413,14 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * temporary file directories.
 	 */
 	public String[] getDbStoragePaths() {
-		if (localRocksDbDirectories == null) {
-			return null;
-		} else {
-			String[] paths = new String[localRocksDbDirectories.length];
-			for (int i = 0; i < paths.length; i++) {
-				paths[i] = localRocksDbDirectories[i].toString();
-			}
-			return paths;
-		}
+		return rocksDBStateBackend.getDbStoragePaths();
 	}
 
 	/**
 	 * Gets whether incremental checkpoints are enabled for this state backend.
 	 */
 	public boolean isIncrementalCheckpointsEnabled() {
-		return enableIncrementalCheckpointing.getOrDefault(CheckpointingOptions.INCREMENTAL_CHECKPOINTS.defaultValue());
+		return rocksDBStateBackend.isIncrementalCheckpointsEnabled();
 	}
 
 	/**
@@ -764,14 +428,14 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @return The type of the priority queue state.
 	 */
 	public PriorityQueueStateType getPriorityQueueStateType() {
-		return priorityQueueStateType == null ? TIMER_SERVICE_FACTORY.defaultValue() : priorityQueueStateType;
+		return LegacyEnumBridge.convert(rocksDBStateBackend.getPriorityQueueStateType());
 	}
 
 	/**
 	 * Sets the type of the priority queue state. It will fallback to the default value, if it is not explicitly set.
 	 */
 	public void setPriorityQueueStateType(PriorityQueueStateType priorityQueueStateType) {
-		this.priorityQueueStateType = checkNotNull(priorityQueueStateType);
+		rocksDBStateBackend.setPriorityQueueStateType(LegacyEnumBridge.convert(priorityQueueStateType));
 	}
 
 	// ------------------------------------------------------------------------
@@ -789,7 +453,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @param options The options to set (must not be null).
 	 */
 	public void setPredefinedOptions(@Nonnull PredefinedOptions options) {
-		predefinedOptions = checkNotNull(options);
+		rocksDBStateBackend.setPredefinedOptions(options);
 	}
 
 	/**
@@ -805,10 +469,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 */
 	@VisibleForTesting
 	public PredefinedOptions getPredefinedOptions() {
-		if (predefinedOptions == null) {
-			predefinedOptions = PredefinedOptions.DEFAULT;
-		}
-		return predefinedOptions;
+		return rocksDBStateBackend.getPredefinedOptions();
 	}
 
 	/**
@@ -825,7 +486,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @param optionsFactory The options factory that lazily creates the RocksDB options.
 	 */
 	public void setRocksDBOptions(RocksDBOptionsFactory optionsFactory) {
-		this.rocksDbOptionsFactory = optionsFactory;
+		rocksDBStateBackend.setRocksDBOptions(optionsFactory);
 	}
 
 	/**
@@ -838,15 +499,14 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 */
 	@Nullable
 	public RocksDBOptionsFactory getRocksDBOptions() {
-		return rocksDbOptionsFactory;
+		return rocksDBStateBackend.getRocksDBOptions();
 	}
 
 	/**
 	 * Gets the number of threads used to transfer files while snapshotting/restoring.
 	 */
 	public int getNumberOfTransferThreads() {
-		return numberOfTransferThreads == UNDEFINED_NUMBER_OF_TRANSFER_THREADS ?
-			CHECKPOINT_TRANSFER_THREAD_NUM.defaultValue() : numberOfTransferThreads;
+		return rocksDBStateBackend.getNumberOfTransferThreads();
 	}
 
 	/**
@@ -855,9 +515,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @param numberOfTransferThreads The number of threads used to transfer files while snapshotting/restoring.
 	 */
 	public void setNumberOfTransferThreads(int numberOfTransferThreads) {
-		Preconditions.checkArgument(numberOfTransferThreads > 0,
-			"The number of threads used to transfer files in RocksDBStateBackend should be greater than zero.");
-		this.numberOfTransferThreads = numberOfTransferThreads;
+		rocksDBStateBackend.setNumberOfTransferThreads(numberOfTransferThreads);
 	}
 
 	/**
@@ -880,8 +538,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * Gets the max batch size will be used in {@link RocksDBWriteBatchWrapper}.
 	 */
 	public long getWriteBatchSize() {
-		return writeBatchSize == UNDEFINED_WRITE_BATCH_SIZE ?
-			WRITE_BATCH_SIZE.defaultValue().getBytes() : writeBatchSize;
+		return rocksDBStateBackend.getWriteBatchSize();
 	}
 
 	/**
@@ -890,8 +547,7 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 	 * @param writeBatchSize The size will used to be used in {@link RocksDBWriteBatchWrapper}.
 	 */
 	public void setWriteBatchSize(long writeBatchSize) {
-		checkArgument(writeBatchSize >= 0, "Write batch size have to be no negative.");
-		this.writeBatchSize = writeBatchSize;
+		rocksDBStateBackend.setWriteBatchSize(writeBatchSize);
 	}
 
 	// ------------------------------------------------------------------------
@@ -900,27 +556,17 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 
 	@VisibleForTesting
 	RocksDBResourceContainer createOptionsAndResourceContainer() {
-		return createOptionsAndResourceContainer(null);
-	}
-
-	@VisibleForTesting
-	private RocksDBResourceContainer createOptionsAndResourceContainer(
-		@Nullable OpaqueMemoryResource<RocksDBSharedResources> sharedResources) {
-
-		return new RocksDBResourceContainer(
-			predefinedOptions != null ? predefinedOptions : PredefinedOptions.DEFAULT,
-			rocksDbOptionsFactory,
-			sharedResources);
+		return rocksDBStateBackend.createOptionsAndResourceContainer();
 	}
 
 	@Override
 	public String toString() {
 		return "RocksDBStateBackend{" +
 				"checkpointStreamBackend=" + checkpointStreamBackend +
-				", localRocksDbDirectories=" + Arrays.toString(localRocksDbDirectories) +
-				", enableIncrementalCheckpointing=" + enableIncrementalCheckpointing +
-				", numberOfTransferThreads=" + numberOfTransferThreads +
-				", writeBatchSize=" + writeBatchSize +
+				", localRocksDbDirectories=" + Arrays.toString(rocksDBStateBackend.getDbStoragePaths()) +
+				", enableIncrementalCheckpointing=" + rocksDBStateBackend.isIncrementalCheckpointsEnabled() +
+				", numberOfTransferThreads=" + rocksDBStateBackend.getNumberOfTransferThreads() +
+				", writeBatchSize=" + rocksDBStateBackend.getWriteBatchSize() +
 				'}';
 	}
 
@@ -930,68 +576,11 @@ public class RocksDBStateBackend extends AbstractManagedMemoryStateBackend
 
 	@VisibleForTesting
 	static void ensureRocksDBIsLoaded(String tempDirectory) throws IOException {
-		synchronized (RocksDBStateBackend.class) {
-			if (!rocksDbInitialized) {
-
-				final File tempDirParent = new File(tempDirectory).getAbsoluteFile();
-				LOG.info("Attempting to load RocksDB native library and store it under '{}'", tempDirParent);
-
-				Throwable lastException = null;
-				for (int attempt = 1; attempt <= ROCKSDB_LIB_LOADING_ATTEMPTS; attempt++) {
-					File rocksLibFolder = null;
-					try {
-						// when multiple instances of this class and RocksDB exist in different
-						// class loaders, then we can see the following exception:
-						// "java.lang.UnsatisfiedLinkError: Native Library /path/to/temp/dir/librocksdbjni-linux64.so
-						// already loaded in another class loader"
-
-						// to avoid that, we need to add a random element to the library file path
-						// (I know, seems like an unnecessary hack, since the JVM obviously can handle multiple
-						//  instances of the same JNI library being loaded in different class loaders, but
-						//  apparently not when coming from the same file path, so there we go)
-
-						rocksLibFolder = new File(tempDirParent, "rocksdb-lib-" + new AbstractID());
-
-						// make sure the temp path exists
-						LOG.debug("Attempting to create RocksDB native library folder {}", rocksLibFolder);
-						// noinspection ResultOfMethodCallIgnored
-						rocksLibFolder.mkdirs();
-
-						// explicitly load the JNI dependency if it has not been loaded before
-						NativeLibraryLoader.getInstance().loadLibrary(rocksLibFolder.getAbsolutePath());
-
-						// this initialization here should validate that the loading succeeded
-						RocksDB.loadLibrary();
-
-						// seems to have worked
-						LOG.info("Successfully loaded RocksDB native library");
-						rocksDbInitialized = true;
-						return;
-					}
-					catch (Throwable t) {
-						lastException = t;
-						LOG.debug("RocksDB JNI library loading attempt {} failed", attempt, t);
-
-						// try to force RocksDB to attempt reloading the library
-						try {
-							resetRocksDBLoadedFlag();
-						} catch (Throwable tt) {
-							LOG.debug("Failed to reset 'initialized' flag in RocksDB native code loader", tt);
-						}
-
-						FileUtils.deleteDirectoryQuietly(rocksLibFolder);
-					}
-				}
-
-				throw new IOException("Could not load the native RocksDB library", lastException);
-			}
-		}
+		EmbeddedRocksDBStateBackend.ensureRocksDBIsLoaded(tempDirectory, LOG);
 	}
 
 	@VisibleForTesting
 	static void resetRocksDBLoadedFlag() throws Exception {
-		final Field initField = org.rocksdb.NativeLibraryLoader.class.getDeclaredField("initialized");
-		initField.setAccessible(true);
-		initField.setBoolean(null, false);
+		EmbeddedRocksDBStateBackend.resetRocksDBLoadedFlag();
 	}
 }
