+++ flink-state-backends/flink-statebackend-rocksdb/src/test/java/org/apache/flink/contrib/streaming/state/RocksDBInitTest.java
@@ -29,6 +29,8 @@ import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.rocksdb.RocksDB;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
@@ -42,6 +44,8 @@ import static org.junit.Assert.fail;
 @PrepareForTest({RocksDB.class})
 public class RocksDBInitTest {
 
+	private static final Logger LOG = LoggerFactory.getLogger(RocksDBInitTest.class);
+
 	@Rule
 	public final TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -50,7 +54,7 @@ public class RocksDBInitTest {
 	 */
 	@Test
 	public void testResetInitFlag() throws Exception {
-		RocksDBStateBackend.resetRocksDBLoadedFlag();
+		EmbeddedRocksDBStateBackend.resetRocksDBLoadedFlag();
 	}
 
 	@Test
@@ -60,7 +64,7 @@ public class RocksDBInitTest {
 
 		File tempFolder = temporaryFolder.newFolder();
 		try {
-			RocksDBStateBackend.ensureRocksDBIsLoaded(tempFolder.getAbsolutePath());
+			EmbeddedRocksDBStateBackend.ensureRocksDBIsLoaded(tempFolder.getAbsolutePath(), LOG);
 			fail("Not throwing expected exception.");
 		} catch (IOException ignored) {
 			// ignored
