+++ flink-state-backends/flink-statebackend-rocksdb/src/test/java/org/apache/flink/contrib/streaming/state/RocksDBTestUtils.java
@@ -62,7 +62,7 @@ public final class RocksDBTestUtils {
 			new KeyGroupRange(0, 1),
 			new ExecutionConfig(),
 			TestLocalRecoveryConfig.disabled(),
-			RocksDBStateBackend.PriorityQueueStateType.HEAP,
+			EmbeddedRocksDBStateBackend.PriorityQueueStateType.HEAP,
 			TtlTimeProvider.DEFAULT,
 			new UnregisteredMetricsGroup(),
 			Collections.emptyList(),
@@ -91,7 +91,7 @@ public final class RocksDBTestUtils {
 				new KeyGroupRange(0, 1),
 				new ExecutionConfig(),
 				TestLocalRecoveryConfig.disabled(),
-				RocksDBStateBackend.PriorityQueueStateType.HEAP,
+				EmbeddedRocksDBStateBackend.PriorityQueueStateType.HEAP,
 				TtlTimeProvider.DEFAULT,
 				new UnregisteredMetricsGroup(),
 				Collections.emptyList(),
